feat:add node&task api,along with tests

+++ api/sandglass_api/db.py
@@ -8,6 +8,8 @@ def connect_to(db_uri: str, db_name):
 
 # WARNING:THIS WILL DROP **ALL** COLLECTIONS
 # DO NOT USE IN PRODUCTION ENV
-def RESET_DATABASE():
+def RESET_DATABASE(config):
+    if not config.get("TESTING"):
+        raise Exception("RESET_DATABASE should only be used in testing environment")
     for t in ALL_DOCUMENTED_TYPE:
         t.drop_collection()
